+++ src/ll/pysql.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 # cython: language_level=3, always_allow_keywords=True
 
-## Copyright 2012-2020 by LivingLogic AG, Bayreuth/Germany
-## Copyright 2012-2020 by Walter Dörwald
+## Copyright 2012-2021 by LivingLogic AG, Bayreuth/Germany
+## Copyright 2012-2021 by Walter Dörwald
 ##
 ## All Rights Reserved
 ##