+++ test/test_pysql.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 # cython: language_level=3, always_allow_keywords=True
 
-## Copyright 2014-2020 by LivingLogic AG, Bayreuth/Germany
-## Copyright 2014-2020 by Walter Dörwald
+## Copyright 2014-2021 by LivingLogic AG, Bayreuth/Germany
+## Copyright 2014-2021 by Walter Dörwald
 ##
 ## All Rights Reserved
 ##